Add files via upload
This dissertation is based on studying the physiological properties of the plants, considering the external properties of the plants, like soil moisture level, amount of sunlight received, and humidity in the environment.
This cost-effective approach provides a strong incentive for further exploration of plant biophysics and bioacoustics. By observing and analyzing external signals and frequencies, we can gain a deeper understanding of these phenomena through the lens of physics.